third_party/blink/web_tests/platform/generic/http/tests/websocket/null-character-expected.txt

Make sure WebSocket transfer null character

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

WebSocket open
msg should not be split by frame char \xff\0
PASS msg is expectedMessage
WebSocket closed
PASS successfullyParsed is true

TEST COMPLETE
